dts: galen: add card name for HDA
We have an unified audio XML file for android and HDMI port on galen
uses different hda pcm device. This causes conflicts with the existing
node for hda. Hence a different card name will be maintained for galen,
the same is passed with current patch.

Bug 200375020

Change-Id: Ia06a43585717dae06fd35af86f9216aa894d28c2
Reviewed-on: https://git-master.nvidia.com/r/1626533
Tested-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>