gpio: t8x: Set direction of GPIO in irq_unmask
There is client of GPIOs which get the interrupt of GPIO
from the  DT directly. On this case, they call the irq_*
APIs for the enabling/disabling interrupt.

Make sure that GPIO pins are in correct configurations on
such cases. For such cases, making pin in gpio input
direction

Bug 2039546

Change-Id: Ie63e1fa9760a197753cc81281f29bbe1ddb0b636
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1627185
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>