cpufreq: t194: return promptly in case of failure
If nvfreq_feedback returns same counters in consecutive reads,
return -EIO immediately. This helps functions deep in the call
stack handle the failure properly.

For reading freq through debugfs, return 0 if CPU is not online.
This reduces confusion when a large random number is returned to
caller. Similarly, for reading ndiv from debugfs, return 0 if
CPU is not online. the minimum ndiv is larger than 0, so there
is no confusion here.

Change-Id: I387f56743e01117994798cbb2e9ada1545fbcff7
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1628581
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>