ethernet: eqos: disable USO
EQOS in Xavier supports UDP Segmentation Offload (USO)
in which the UDP payload is segmented in the hardware.

In case of USO. engine updates only the following fields
in the IP-UDP headers:
1) Total Length, Identification Field, and Header Checksum
   field in IPv4 header
2) Payload Length field in IPv6 header
3) Length and Checksum field in UDP header

There are no sequencing/ordering controls available or
updated in the segments, so it can lead to out of order
packets on the receiver side.

Reverting USO feature for fixing out of order packets issue.

Bug 200379677

Change-Id: I95afa8ceb55d20fa759890de74ccb59b31e24af2
Signed-off-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1653555
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>