nvlink: Update t19x nvlink endpoint IOCTL API
This change updates t19x nvlink endpoint IOCTL API in
following ways:

1. Use kzalloc() instead if devm_zalloc() as dev_zalloc()
seems to be causing issues in MODs environment when IOCTLs
are stress tested.

2. Check if argument size received to IOCTL api is matches
with that of the structure size.

3. Do copy_from_user() at the top of API and copy_to_user()
at the bottom of API, instead of doing them individually for
each commands passed to ioctl API.

Change-Id: If9fa4cff3b7c2b887fca0a6d9830c031d5f9d097
Signed-off-by: Petlozu Pravareshwar <petlozup@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1631071
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vijayakumar Subbu <vsubbu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>