video: tegra: host: Skip GoS if ctxs are isolated
Currently GoS is enabled always for the devices that support GoS,
however, this is not required in cases where engine channel submission
mode is enabled: All waits in this scneario are done on Host1x and the
engine does not need to access the GoS or syncpoint<->MSS interface.

Further, if context isolation is enabled as part of using channel
submission mode, the GoS mappings are not available thereby causing
invalid memory accesses if attempted.

This change modifies GoS code to skip over allocation of syncpoint
GoS backing if context isolation is enabled for a device.

Bug 2007922

Change-Id: I7b0eaf6e5fea1fe2266220682d0fee27674405a4
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1643630
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-by: Deepak Nibade <dnibade@nvidia.com>
Reviewed-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>