tegra: camera: rtcpu: reboot after failed tests
Flush the RTCPU tracer and print out the RTCPU log entries after a
test run.

Reboot rtcpu after a failed test run. An unsuccessful test test may
leave RTCPU in a bad state with references to the soon-to-be-unmapped
memory, which may crash the whole device.

Bug 2036751

Change-Id: I9dbabfbdbb88d9e0afd58451f8d2f25e6d263282
Signed-off-by: Pekka Pessi <ppessi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1653628
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>