hdmi: fix PLLD* programming for RGB/YUV444 12bpc
For RGB/YUV444 12bpc, the pclk:orclk needs to be in a
2:3 ratio. For example, if we're driving 4k@30 YUV444
12bpc with a pclk of 297MHz, the orclk (TMDS char rate)
needs to be running at 297MHz * 1.5 = 445.5MHz.

If the SOR brick is running in HDMI1.4 mode (TMDS <= 340),
the SOR reference clk should be running at the TMDS rate.
The SOR reference clk vs. pixel clk dividers should be in
a 2:3 ratio in this mode.

If the SOR brick is running in HDMI2.0 mode (TMDS > 340),
the SOR reference clk needs to be divided down by 2, and
the SOR link multiplier should be programmed to 20x (instead
of 10x) to account for this. The SOR reference clk vs. pixel clk
dividers should be in a 4:3 ratio in this mode.

In order to satisfy the above constraints, the PLLD* source
that we're using should be set to PCLK * 3. The max pclk
that we need to practically support for deepcolor modes
is 400MHz (600MHz TMDS rate), so PCLK * 3 should never
exceed the max frequency that we can hit on the PLLD*
clks. We're currently using PCLK * 6, which is way too high.
For example, this would cause us to exceed the max frequency on
PLLD2/PLLD3/PLLD4 for 4k@30.

Bug 2043009

Change-Id: I2191dd87637463b1dbca2e000f23c1ac5c43dd78
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1647004
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Santosh Galma <galmar@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>