arm64: configs: tegra: Disable TRANSPARENT_HUGEPAGE
The kernel is hitting a BUG at:
    kernel BUG at /build/ddastoussthi/N/kernel/kernel-4.9/mm/gup.c:718!

This BUG_ON() is supposed to catch any unhandled replaceable CMA page.
However, the code doesn't handle transparent huge pages at all as it
was not enabled in K4.4.

So this patch temporarily disables the CONFIG_TRANSPARENT_HUGEPAGE till
the code has the migration support for huge pages as well.

Bug 2038622

Change-Id: If9dbe503101292e98e9bbced2e82f03410ddef93
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1650585
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>