mttcan: Read PSR register in irq handler
We read PSR register in softirq context in case IR register reports that
we have error in either arbitration or data phase. But in case of auto
retransmission mode (which is default), by the time we read PSR register
error status might have changed because of successful transmission.
Correct way is to read the PSR register in irq handler itself in case of
error and process it in softirq.

Bug 200380726
Bug 200383529

Change-Id: Ib696d374c1d740f1949a0087fb6a6e13c00fdabb
Signed-off-by: Abhijit <abhijit@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1652053
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vipin Kumar <vipink@nvidia.com>
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>