mm: do not restore page refcount on failure
According to the original commit 246c7fc012a1 ("mm: introduce
migrate_replace_page() for migrating page to the given target"),
the migrate_replace_page() function was designed in a way which
requires its caller to hold an additional refcount by calling
get_page() before entering the migrate_replace_page() function
and releases the refcount inside the migrate_replace_page() on
success, or remains the refcount on failure:
  caller() {
      ......
      get_page(page);
      migrate_replace_page(page, ....) {
          ....
          put_page(page);
          ret = __migrate(page, ....);
          if (ret)
              get_page(page);
      }
      ......
  }

On the other hand, our downstream design of migrating CMA pages
via this function does not call an additional get_page() and it
removed the put_page() to keep the refcount balanced. However,
it didn't remove the get_page() on failure:
  migrate_replace_cma_page() {
      ......
      migrate_replace_page(page, ....) {
          ....
          ret = __migrate(page, ....);
          if (ret)
              get_page(page);    // refcount increased here
      }
      ......
  }

So this patch removes this get_page() to remain the refcount as
the same number before entering the function.

Bug 200378821

Change-Id: I54a62cc09351e9656d417ef8ddf2a6a4854e93d5
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1652927
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>